requirements.txt
streamlit>=1.41,<2.0
boto3>=1.35,<1.37
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0 